# Reduces allocations during instantiation