… internal/engine/compiler/engine_test.go
… internal/integration_test/bench/hostfunc_bench_test.go
… internal/wasm/binary/decoder_test.go
… internal/wasm/counts_test.go
… internal/wasm/function_definition_test.go
… internal/wasm/module_test.go
… internal/wasm/store_test.go
… internal/wasm/table.go
… internal/wasm/table_test.go
… internal/wazeroir/compiler.go
… runtime.go