~This is a relatively large refactoring, so maybe good to go after 1.0 release~

```
$ benchstat old.txt new2.txt
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: AMD Ryzen 9 3950X 16-Core Processor
                                       │   old.txt   │              new2.txt              │
                                       │   sec/op    │   sec/op     vs base               │
Initialization/interpreter-32            36.10µ ± 1%   34.75µ ± 1%  -3.74% (p=0.000 n=30)
Initialization/interpreter-multiple-32   38.46µ ± 3%   38.39µ ± 2%       ~ (p=0.352 n=30)
Initialization/compiler-32               27.53µ ± 3%   26.07µ ± 7%  -5.32% (p=0.004 n=30)
Initialization/compiler-multiple-32      24.35µ ± 4%   23.76µ ± 4%       ~ (p=0.104 n=30)
Compilation/with_extern_cache-32         197.9µ ± 0%   197.7µ ± 0%       ~ (p=0.924 n=30)
Compilation/without_extern_cache-32      5.875m ± 0%   5.878m ± 0%       ~ (p=0.982 n=30)
geomean                                  101.3µ        99.32µ       -1.98%

                                       │   old.txt    │              new2.txt               │
                                       │     B/op     │     B/op      vs base               │
Initialization/interpreter-32            133.0Ki ± 0%   132.9Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/interpreter-multiple-32   133.8Ki ± 0%   133.8Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/compiler-32               133.6Ki ± 0%   133.6Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/compiler-multiple-32      138.3Ki ± 0%   138.3Ki ± 0%  -0.02% (p=0.000 n=30)
Compilation/with_extern_cache-32         53.33Ki ± 0%   53.33Ki ± 0%  +0.00% (p=0.040 n=30)
Compilation/without_extern_cache-32      1.898Mi ± 0%   1.898Mi ± 0%       ~ (p=0.433 n=30)
geomean                                  180.1Ki        180.1Ki       -0.01%

                                       │   old.txt   │              new2.txt              │
                                       │  allocs/op  │  allocs/op   vs base               │
Initialization/interpreter-32             35.00 ± 0%    32.00 ± 0%  -8.57% (p=0.000 n=30)
Initialization/interpreter-multiple-32    54.00 ± 0%    51.00 ± 0%  -5.56% (p=0.000 n=30)
Initialization/compiler-32                35.00 ± 0%    32.00 ± 0%  -8.57% (p=0.000 n=30)
Initialization/compiler-multiple-32       44.00 ± 0%    41.00 ± 0%  -6.82% (p=0.000 n=30)
Compilation/with_extern_cache-32          982.0 ± 0%    982.0 ± 0%       ~ (p=0.570 n=30)
Compilation/without_extern_cache-32      32.62k ± 0%   32.62k ± 0%       ~ (p=0.580 n=30)
geomean                                   212.9         202.3       -4.99%


```