@mathetake 2023-03-22
looks really promising 

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: AMD Ryzen 9 3950X 16-Core Processor
                                       │   old.txt   │              new.txt               │
                                       │   sec/op    │   sec/op     vs base               │
Initialization/interpreter-32            36.10µ ± 1%   34.86µ ± 1%  -3.43% (p=0.000 n=30)
Initialization/interpreter-multiple-32   38.46µ ± 3%   37.90µ ± 4%       ~ (p=0.297 n=30)
Initialization/compiler-32               27.53µ ± 3%   27.98µ ± 4%       ~ (p=0.965 n=30)
Initialization/compiler-multiple-32      24.35µ ± 4%   23.89µ ± 2%       ~ (p=0.149 n=30)
Compilation/with_extern_cache-32         197.9µ ± 0%   196.3µ ± 0%  -0.84% (p=0.000 n=30)
Compilation/without_extern_cache-32      5.875m ± 0%   5.846m ± 0%  -0.49% (p=0.008 n=30)
geomean                                  101.3µ        100.2µ       -1.09%

                                       │   old.txt    │               new.txt               │
                                       │     B/op     │     B/op      vs base               │
Initialization/interpreter-32            133.0Ki ± 0%   132.9Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/interpreter-multiple-32   133.8Ki ± 0%   133.8Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/compiler-32               133.6Ki ± 0%   133.6Ki ± 0%  -0.02% (p=0.000 n=30)
Initialization/compiler-multiple-32      138.3Ki ± 0%   138.3Ki ± 0%  -0.02% (p=0.000 n=30)
Compilation/with_extern_cache-32         53.33Ki ± 0%   53.33Ki ± 0%       ~ (p=0.194 n=30)
Compilation/without_extern_cache-32      1.898Mi ± 0%   1.898Mi ± 0%  -0.00% (p=0.044 n=30)
geomean                                  180.1Ki        180.1Ki       -0.02%

                                       │   old.txt   │              new.txt               │
                                       │  allocs/op  │  allocs/op   vs base               │
Initialization/interpreter-32             35.00 ± 0%    32.00 ± 0%  -8.57% (p=0.000 n=30)
Initialization/interpreter-multiple-32    54.00 ± 0%    51.00 ± 0%  -5.56% (p=0.000 n=30)
Initialization/compiler-32                35.00 ± 0%    32.00 ± 0%  -8.57% (p=0.000 n=30)
Initialization/compiler-multiple-32       44.00 ± 0%    41.00 ± 0%  -6.82% (p=0.000 n=30)
Compilation/with_extern_cache-32          982.0 ± 0%    982.0 ± 0%       ~ (p=0.334 n=30)
Compilation/without_extern_cache-32      32.62k ± 0%   32.62k ± 0%       ~ (p=0.230 n=30)
geomean                                   212.9         202.3       -4.99%

```