@codefromthecrypt 2023-03-22
personally, I'm ok on this with some fuzz time, assuming our fuzzer hits tables etc.